updated to take a floating point for input

+++ mphfps-v0p.py
@@ -1,3 +1,3 @@
 n = input('Enter MPH - ')
-print((int(n)*66)/45,'feet per second')
+print((float(n)*66)/45,'feet per second')
